Optimize UidSet TakeRange.
Once again, let's hope the unit tests are exhaustive enough...
Also, Contains seems to be a bottleneck in unit tests, though I doubt
that will be the case in real performance. It would probably be worth-
while to implement a skiplist for this either way (linked list nodes
that also point e.g. 10 or 100 elements ahead) as a bandaid fix.